update redux and component directories
