+++ src/App.js
@@ -2,9 +2,8 @@
 
 import './App.css';
 import { Route, Routes } from 'react-router-dom';
-import BookContainer from './redux/components/BookContainer';
+import BookContainer from './components/BookContainer';
 import CategoriesPage from './components/CategoriesPage';
-// import HomePage from './components/HomePage';
 import Header from './components/Header';
 
 function App() {
